fix compile error

+++ solidity/contracts/BellTower.sol
@@ -1,19 +1,16 @@
+// SPDX-License-Identifier: Unlicensed
 pragma solidity ^0.8.4;
 
-
 contract BellTower {
-
     // counter for how many times the bell has benn rung
-    uint bellRung;
+    uint256 bellRung;
 
     // event for ringing a bellRung
-    event BellRung(uint rungForTheTime, address whoRungIt);
+    event BellRung(uint256 rungForTheTime, address whoRungIt);
 
     function ringBell() public {
-        bellRung ++;
+        bellRung++;
 
         emit BellRung(bellRung, msg.sender);
     }
-
 }
-
